更多统计字段展示

+++ src/main/java/net/csibio/propro/algorithm/lfqbench/bean/ProteinRatio.java
@@ -1,5 +1,5 @@
 package net.csibio.propro.algorithm.lfqbench.bean;
 
-public record ProteinRatio(String protein, float x, float y) {
+public record ProteinRatio(String protein, double x, double y) {
 
 }
